Moved service client sendasync timeout test to different class.
